[BUGFIX] Fatal error when exporting from root
An uncaught exception happens when trying to export records from a
singe table view in List mode on the page tree root. This is due to
the tree root being handled as a normal page record.

With this patch the tree root is handled as it should. Some cleanup
(CGL) is also applied.

Change-Id: Ia5c46044f5040472846f3e7fb4adbf203e5e1e6a
Fixes: #31350
Releases: 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/6551
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Georg Ringer
Reviewed-by: Steffen Ritter
Reviewed-by: Francois Suter
Tested-by: Francois Suter