[TASK] Limit the exported record fields in impexp
To be able to compare mostly only the data provided in the
database fixtures, the export needs to be limited to a given
set of fields of a record type.
This will loose the strict xml comparison of all existing fields
in the database and the tests would not break whenever we
change ext:core or ext:frontend ext_tables.sql.

Resolves: #55625
Releases: 6.2
Change-Id: Ie48e540ab1c9e271b64b893353a84c34ac982a50
Reviewed-on: https://review.typo3.org/27303
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Tymoteusz Motylewski
Reviewed-by: Marc Bastian Heinrichs
Tested-by: Marc Bastian Heinrichs