[BUGFIX] Use created folder object for adding images in impexp
A folder to be created could change the name by sanitizing. So
the name of the created folder could differ from the requested
name. For adding the image the returned folder object should be
used.

To avoid that sanitized ones should be created over and over again
a mapping is buffered.

Resolves: #58682
Releases: 6.2
Change-Id: Ia3443c36c6dda8531c8791c221d2a835b25b10b2
Reviewed-on: https://review.typo3.org/29984
Reviewed-by: Bernd Niehues
Tested-by: Bernd Niehues
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters