[TASK] Extract flex form language handling to compatibility6
Flex form has its own localization handling that exists in parallel
to the record overlay based system. This was introduced together with
the TemplaVoila concept and is almost exclusively used in this context.
For details, the logic is described in the flex form section of
the "localization guide".

With the FormEngine refactoring it is now possible to extract most
parts of this language handling to ext:compatibility6. The patch
moves the current handling classes to compatibility6 and registers
them.

This extraction gives TemplaVoila and maybe other extensions the
opportunity to implement this handling on their own, compatibility6
can be used as a show case how that may be achieved.

As main benefit the classes in FormEngine are significantly
shorter, easier to understand and quite a few options can
be dropped.

Resolves: #70138
Releases: master
Change-Id: Ia2c4b8b329702b895b9183f835423c2d27b7c74a
Reviewed-on: http://review.typo3.org/43601
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>