[!!!][TASK] Remove database field pages.alias
This is a precursor for removing PseudoSiteHandling in general.

The database field "pages.alias" field is dropped, along with
the functionality to evalute if a frontend request "?id=acme"
is non-integer, as it now always has to be integer.

Existing links pointing to page aliases will stop working.

Resolves: #87356
Releases: master
Change-Id: I19134cc788e633e140b43497f716082ac96744e5
Reviewed-on: https://review.typo3.org/59232
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>