[TASK] Extract "taskcenter" and "sys_action" to TER/composer
This is the last round of planned extension extractions
from the main core repository.

Both the extensions 'taskcenter' and its add-on extension
'sys_action' did not receive much love over the years. They
are rarely used and their approach is clumsy to say the least.

They are both dropped from main core repository with core v10
now. A fresh approach for a cool new dashboard is in the works
already that may eventually find its way into the core later.

The patch removes both extensions with have been released as
'taskcenter' and 'sys_action' to TER and as
'friendsoftypo3/taskcenter' and 'friendsoftypo3/sys_action'
to composer.

Extension 'impexp' registered a task for the taskcenter which
is now within 'taskcenter' directly.

Resolves: #87499
Releases: master
Change-Id: I86e0292f13247bbd1f2060214f4f6d8e2aefe6c0
Reviewed-on: https://review.typo3.org/59502
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>