[!!!][TASK] FAL: Decouple logic of ResourceFactory into StorageRepository
The "ResourceFactory" class was introduced as a "god class" to build
any File/Folder object (from a string, DB row or integer), but is also
there for building sys_file_storage (ResourceStorage) objects, so it is
mixing a lot of concerns (also driver setup).

However, all creation of ResourceStorage objects should be moved into
the StorageRepository (which already exists), thus reducing
the dependencies towards ResourceFactory.

This makes it much easier to not juggle with storage UIDs anymore
but actually work with storage objects throughout the internal FAL
API.

The StorageRepository is not extending from the AbstractRepository
anymore, which is breaking.

Resolves: #92289
Releases: master
Change-Id: Ib367164535550e280a6907f4d4473aa0417c3813
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65184
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>