[BUGFIX] Register icons not being part of TYPO3.Icons again
Since #92689 the IconRegistry is fed by a JSON file taken from the
TYPO3.Icons repository and does not grab all PNG and SVG files from the
filesystem and auto-registers these files anymore.

It turned out that some icons (all of them being PNG files) are not part
of the TYPO3.Icons repository and thus haven't been registered anymore.

This patch now registers icons of EXT:backend and EXT:impexp explicitly
which are not part of the beforehand mentioned repository.

Resolves: #92860
Related: #92689
Releases: master, 10.4
Change-Id: I0baab2b00be100ad768b4a67bf678e71e11f70bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66693
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>