[FEATURE] Implement lazy console command list
Based on the configuration syntax of the Symfony Console
feature https://github.com/symfony/symfony/pull/39851
…but implemented differently, using a registry pattern
rather then a lazy-object pattern (like symfony does).

Main motiviation for the registry pattern is following:
Symfony LazyCommand wrappers add quite some complexity only
for the sake of the list command, we already got lazy
commands (in terms of execution) as our CommandRegistry
implements the ConfigurationLoaderInterface that has been
introduced by 2017 to add support for lazy commands.

Now, that means we already got a registry for lazy commands,
so it is logical to add lazy description handling there as well.

We want to assure that the command list will never instantiate
any commands. This is in constrast to the Symfony core LazyCommand
approach, where legacy commands, that do not provide a compile time
description, would still be instantiated during console command list.

Also commands that return false in `isEnabled()` are now listed.
That means enabled state is only evaluated during runtime.
Therefore the special `dumpautoload` command is transformed into a
lowlevel command in order to be hidden dependending on being run
in composer-mode or not.

Releases: master
Resolves: #93174
Change-Id: Ifa68404cc81c64a335be30f2263a7eb17de0624d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67635
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Helmut Hummel <typo3@helhum.io>