[BUGFIX] Enable Export/Import of page translations
The impexp system extension is now able to export translated pages. This
essential functionality has not been implemented up until now.

Also, specific language variants can be exported, that will only
contain records matching the requested language. This "hidden
feature" can be accessed by right-clicking on a page's icon in the
translation overview and selecting "export". This feature was
available before, but the export never included actual translated
records before.

In addition, the relation field "l10n_source" is taken care of while
exporting/importing. Normally this would happen automatically, if there
was a refindex entry for that. However, no references are saved for
the tables "pages" and "tt_content". This seems to be another bug and
should be fixed in another patch. After this, some code can be cleaned
up.

Functional tests have been added to verify the exported data.

Resolves: #93470
Resolves: #94616
Releases: main, 12.4
Change-Id: I6f8ab179c524b161dd7cf083cdc89b8cc05bf8b0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71374
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>