[BUGFIX] Suppress E_DEPRECATED for strftime() and gmstrftime()
PHP functions strftime() and gmstrftime() have been deprecated in
PHP 8.1 and will be removed in PHP 9.0. Currently, there is no way
to replace this seamlessly as no drop-in replacement is available.

Official PHP 8.1 deprecation RFC:
https://wiki.php.net/rfc/deprecations_php_8_1#strftime_and_gmstrftime

strftime() has been used in several places in the core, and even
concrete wrapper like TypoScript 'stdWrap_strftime' or support in
fluid viewhelper 'f:date'. Besides this, some places used it to
format date or locale datetimes, which may be replaceable.

As we cannot provide drop-in replacement which would be breaking,
or adding a hard requirement of the optional php extension intl,
this patch suppresses deprecation warnings on calls of strftime()
and gmstrftime(), using the '@' error control operator.

Further, corresponding @todo's are added along the way to replace
the occurrences in v12.

Resolves: #95875
Related: #95872
Releases: master
Change-Id: Ieb152ab5822f0d2e720a102873f51f4972e3784c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72055
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>