[TASK] Streamline functions returning $GLOBALS['BE_USER']
Many classes in TYPO3 have a function called
getBackendUser() or getBackendUserAuthentication(). The
function definition is not always equal and some functions
have a return type or doc comment and others do not.

This patch streamlines the functions as following:

* They must always use a return type declaration
* They must not contain a doc comment header

Also, some superfluous override of getBackendUser()
have been removed.

Additionally the parameter type declaration for
$backendUserAuthentication in

* BeforeHistoryRollbackStartEvent
* AfterHistoryRollbackFinishedEvent
* RecordHistoryRollback

have been adapted to be nullable.

Resolves: #96752
Releases: main
Change-Id: Id7eec1bc5d0d602ccab95de75149053f7223a86c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73304
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>