[!!!][TASK] Remove context sensitive help
The context sensitive help (CSH) is not accessible and the information
is often not useful or not updated for a long time.

As the online documentation is now improved a lot, there is a link added
in the main help menu. For short, but important additional information
to explain the usage of a field the `description` field should be used.

Resolves: #97312
Releases: main
Change-Id: Ic0b7df90ef88c53f802bfb254023ba030ff01ee7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74191
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>