[TASK] Deprecate ext_emconf.php clearCacheOnLoad
There is little point in not clearing all caches
when extensions are loaded / unloaded using
the extension manager. It's easier for extension
developers to rely on all caches being cleared,
it's one thing less to think about.

The patch deprecates the clearCacheOnLoad and
clearcacheonload array key in ext_emconf.php
and removes the handling in extension manager.

Resolves: #99019
Releases: main
Change-Id: Id6ff3db02e2ff82813534e1d2a40cbfac0013b8c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76474
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>