[TASK] Ensure short list syntax is used
With #90002, all occurrences of the `list()` function have been migrated
to the short list syntax. Our php-cs-fixer configuration does however
not check the short list syntax usage, so we have one usage of `list()`
in our codebase again.

This change extends the php-cs-fixer configuration to ensure, that
only the short list syntax is used. The new ruleset has been applied
to the codebase and one usage of the `list()` function has been migrated.

Resolves: #101777
Related: #90002
Releases: main, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I30f5822e27a36515c2c77a19089a0754ff6fdcfd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80731
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>