[TASK] Remove leftover upload folder code in BE_USER
Due to the introduction of the "DefaultUploadFolderResolver"
class in v12, a dedicated service to load information about
the Default Upload Folder for a user, some internal
methods in BackendUserAuthentication can be removed.

DI is not used due to complications in the inheritance
of the ImportExport class (which needs to be resolved first).

Resolves: #102500
Releases: main
Change-Id: I4ef4001ff258d2ec09d701df488205e7b208e3dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81949
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>